straigthen out types and customer queries (REL-359)

+++ fdb-relational-core/src/main/java/com/apple/foundationdb/relational/recordlayer/RecordTypeTable.java
@@ -255,8 +255,7 @@ public class RecordTypeTable extends RecordTypeScannable<FDBStoredRecord<Message
                         builder.setField(fd, struct.getFloat(i));
                         break;
                     case Types.INTEGER:
-                        // TODO: Currently passing getLong
-                        builder.setField(fd, struct.getLong(i));
+                        builder.setField(fd, struct.getInt(i));
                         break;
                     case Types.BIGINT:
                         builder.setField(fd, struct.getLong(i));
